Remove v1 files
