--- packages/mural-account-chooser/src/components/account-chooser/types.ts
@@ -1,11 +1,0 @@
-interface AutomaticOptions {
-  email: string;
-  action: 'signin' | 'signup';
-  consentSso?: boolean;
-}
-
-export interface AuthorizeParams {
-  auto?: boolean | AutomaticOptions;
-  signup?: boolean;
-}
-// todo: I don't need this file any more I think.
