+++ packages/mural-account-chooser/src/index.ts
@@ -1,5 +1,5 @@
 import AccountChooser, {
   ACCOUNT_CHOOSER_ACTION,
-} from './components/account-chooser/index';
+} from './components/account-chooser';
 
 export { AccountChooser, ACCOUNT_CHOOSER_ACTION };
